--- src/main/java/io/hhplus/step2/user/domain/application/repository/UserRepository.java
@@ -1,8 +1,0 @@
-package io.hhplus.step2.user.domain.application.repository;
-
-import io.hhplus.step2.user.domain.application.User;
-
-public interface UserRepository {
-
-    User findById(Long userId);
-}
